+++ slim/ext/decorator.py
@@ -1,3 +1,4 @@
+from slim.exception import SlimException
 from ..retcode import RETCODE
 from ..base.view import BaseView, AbstractSQLView
 
@@ -12,7 +13,7 @@ def require_role(role=None):
     return _
 
 
-def request_role(role=None):
+def must_be_role(role=None):
     def _(func):
         async def __(view: AbstractSQLView, *args, **kwargs):
             if role != view.current_request_role:
@@ -26,16 +27,10 @@ async def get_ip(view: BaseView) -> bytes:
     return (await view.get_ip()).packed
 
 
-def get_cooldown_decorator(aioredis_instance, default_unique_id_func=get_ip):
+def get_cooldown_decorator(aioredis_instance: object, default_unique_id_func=get_ip) -> object:
     redis = aioredis_instance
 
     def cooldown(interval_value_or_func, redis_key_template, *, unique_id_func=default_unique_id_func, cd_if_unsuccessed=None):
-        if isinstance(interval_value_or_func, int):
-            interval = interval_value_or_func
-            interval_func = None
-        else:
-            interval_func = interval_value_or_func
-
         def wrapper(func):
             async def myfunc(self: BaseView, *args, **kwargs):
                 # 有可能在刚进入的时候，上一轮已经finish了，那么直接退出
@@ -63,9 +58,15 @@ def get_cooldown_decorator(aioredis_instance, default_unique_id_func=get_ip):
                         # 通过豁免，返回
                         return ret
 
+                    # 检查是否使用间隔函数
+                    if not isinstance(interval_value_or_func, int):
+                        # 如果使用间隔函数，亦不排除直接退出的可能
+                        interval = interval_value_or_func(self, unique_id)
+                        if self.is_finished: return
+                    else:
+                        interval = interval_value_or_func
+
                     # 所有跳过条件都不存在，设置正常的expire并退出
-                    if interval_func:
-                        interval = interval_func(unique_id)
                     await redis.set(key, '1', expire=interval)
                     return ret
             return myfunc